Feature user context

+++ src/utils/Api.js
@@ -13,7 +13,7 @@ import config from './utils'
     return res.json();
   };
 
-  getUser() {
+  getUserInfo() {
     return fetch(`${this._baseUrl}/users/me`, {
       headers: this._headers
     })
@@ -28,7 +28,7 @@ import config from './utils'
     .then(res => this._getResponseData(res))
   };
 
-  editUser(data) {
+  setUserInfo(data) {
     return fetch(`${this._baseUrl}/users/me`, {
       method: 'PATCH',
       headers: this._headers,
@@ -46,7 +46,7 @@ import config from './utils'
       method: 'POST',
       headers: this._headers,
       body: JSON.stringify({
-        name: data.title,
+        name: data.name,
         link: data.link
       })
     })
@@ -61,20 +61,21 @@ import config from './utils'
     .then(res=>this._getResponseData(res))
   };
 
-  addlike(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers
-    })
-    .then(res=>this._getResponseData(res))
-  };
-
-  deleteLike(cardId) {
-    return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers
-    })
-    .then(res=>this._getResponseData(res))
+  changeLikeCardStatus(cardId, isLike) {
+    if(!isLike){
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers
+      })
+      .then(res=>this._getResponseData(res))
+    }else{
+      return fetch(`${this._baseUrl}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers
+      })
+      .then(res=>this._getResponseData(res))
+    }
+    
   };
 
   showAvatar() {
@@ -85,7 +86,7 @@ import config from './utils'
     .then(res=>this._getResponseData(res))
   };
 
-  editAvatar(data) {
+  setUserAvatar(data) {
     return fetch(`${this._baseUrl}/users/me/avatar`, {
       method: 'PATCH',
       headers: this._headers,
@@ -99,5 +100,3 @@ import config from './utils'
 };
 
 export const api = new Api(config);
-//console.log(api)
-//export api;
